Tests reorganization: old things moved to subfolder
