+++ pytest_cases/tests/cases/legacy/advanced/test_parameters.py
@@ -1,7 +1,7 @@
 import pytest
 
-from pytest_cases.tests.cases.example_code import super_function_i_want_to_test
-from pytest_cases.tests.cases.utils import nb_pytest_parameters, get_pytest_param
+from ..example_code import super_function_i_want_to_test
+from ..utils import nb_pytest_parameters, get_pytest_param
 
 from pytest_cases import cases_data, CaseDataGetter, THIS_MODULE, cases_generator
 try:
