+++ pytest_cases/tests/cases/legacy/advanced/test_suite_parametrized_cases.py
