+++ pytest_cases/tests/cases/legacy/advanced/test_suite_shared_dict_cases.py
