+++ pytest_cases/tests/cases/legacy/intermediate/test_shared_cases.py
@@ -1,4 +1,4 @@
-from pytest_cases.tests.cases.example_code import super_function_i_want_to_test, super_function_i_want_to_test2
+from ..example_code import super_function_i_want_to_test, super_function_i_want_to_test2
 
 from pytest_cases import test_target, case_tags
 try:  # python 3.5+
