+++ pytest_cases/tests/cases/legacy/simple/test_generators.py
@@ -1,4 +1,4 @@
-from pytest_cases.tests.cases.example_code import super_function_i_want_to_test
+from ..example_code import super_function_i_want_to_test
 
 from pytest_cases import cases_data, CaseDataGetter, THIS_MODULE, cases_generator
 try:
