+++ pytest_cases/tests/cases/legacy/simple/test_pytest_marks.py
@@ -1,7 +1,7 @@
 import pytest
 
 from pytest_cases.common_pytest import transform_marks_into_decorators
-from pytest_cases.case_parametrizer import get_pytest_marks_on_function, make_marked_parameter_value
+from pytest_cases.case_parametrizer_legacy import get_pytest_marks_on_function, make_marked_parameter_value
 
 
 @pytest.mark.skipif(True, reason="why")
