+++ pytest_cases/tests/cases/legacy/simple/test_so.py
